requirements.txt
scipy>=1.0.0
numpy>=1.16
Keras>=2.2.0
tensorflow_gpu>=1.11.0
dirt
opencv
tqdm
chumpy
pillow
torch
torchvision
einops
keras_layer_normalization

